Refactoring in progress on network accounts
