+++ app/models/Account.java
@@ -1,5 +1,7 @@
 package models;
 
+import java.util.Collection;
+import java.util.List;
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
@@ -7,13 +9,14 @@ import javax.persistence.Enumerated;
 import javax.persistence.Inheritance;
 import javax.persistence.InheritanceType;
 import javax.persistence.ManyToOne;
+import models.activity.StatusActivity;
 import org.apache.commons.lang.builder.EqualsBuilder;
 import org.apache.commons.lang.builder.HashCodeBuilder;
 import play.data.validation.Required;
 import play.db.jpa.Model;
 
 /**
- * An account on an authentication provider
+ * An account on an social network
  * @author Sryl <cyril.lacote@gmail.com>
  */
 @Entity
@@ -29,19 +32,33 @@ public abstract class Account extends Model {
     @Required
     @Enumerated(EnumType.STRING)
     public ProviderType provider; 
+
+    @ManyToOne
+    public OAuthAccount oauthAccount;
+    
+    /** ID of last status retrieved */
+    public String lastStatusId;
     
     public Account(ProviderType provider) {
         this.provider = provider;
     }
     
-    /**
-     * Initialize member profile from account data
-     */
-    public abstract void initMemberProfile();
-    
     public static Account find(ProviderType provider, String login) {
         return Account.find("from Account a where a.provider=?1 and a.member.login=?2", provider, login).first();
     }
+    
+    /**
+     * Fetch recent activities from provider for the given member's account
+     * @param account member's account
+     * @return activities last activities retrieved
+     */
+    public abstract List<StatusActivity> fetchActivities();
+    
+    /**
+     * Enhance given status activities with Link-IT features (ex : link to Link-IT profile, ...)
+     * @param activities status activities to enhance
+     */
+    public abstract void enhance(Collection<StatusActivity> activities);
     
     @Override
     public String toString(){
