+++ app/models/GoogleAccount.java
@@ -1,55 +1,74 @@
 package models;
 
+import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import helpers.JSON;
+import java.text.DateFormat;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Date;
+import java.util.List;
+import java.util.Locale;
 import javax.persistence.Entity;
-import org.apache.commons.lang.StringUtils;
+import models.activity.StatusActivity;
+import play.Logger;
 import play.data.validation.Required;
+import play.libs.WS;
 
 /**
  * A Google account
  * @author Sryl <cyril.lacote@gmail.com>
  */
 @Entity
-public class GoogleAccount extends OAuthAccount {
+public class GoogleAccount extends Account {
     
-    public String googleId;     // 114128610730314333831
     @Required
-    public String email;        // cyril.lacote@gmail.com
-    public String name;         // Cyril Lacôte
-    public String givenName;    // Cyril
-    public String familyName;   // Lacôte
-    public String link;         // http://profiles.google.com/cyrillacote
-    public String picture;      // https://lh5.googleusercontent.com/-ZoXanD5pbxM/AAAAAAAAAAI/AAAAAAAANks/ECGcSElQ6hM/photo.jpg
-    public String gender;       // male
-    public String birthday;     // 0000-03-26 (yes, 0000 for me?!)
-    public String locale;       // en
+    public String googleId;     // 114128610730314333831
+
+    //2011-10-04T14:41:40.837Z
+    static final String DATE_FORMAT = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'";
     
-    public GoogleAccount(String token, String secret) {
-        super(ProviderType.Google, token, secret);
+    public GoogleAccount() {
+        super(ProviderType.Google);
     }
     
     @Override
     public String toString(){
-        return "provider {" + provider + "}, name {" + name + "}";
+        return "Google+ account " + googleId;
     }
 
-    @Override
-    public String getOAuthLogin() {
-        return email;
+    public static GoogleAccount findByEmail(final String email) {
+        return GoogleAccount.find("member.email = ?", email).first();
     }
 
-    @Override
-    public void initMemberProfile() {
-        if (member != null) {
-            if (StringUtils.isBlank(member.googlePlusId)) member.googlePlusId = this.googleId;
-            if (StringUtils.isBlank(member.email)) member.email = this.email;
-            if (StringUtils.isBlank(member.firstname)) member.firstname = this.givenName;
-            if (StringUtils.isBlank(member.lastname)) member.lastname = this.familyName;
-            if (StringUtils.isBlank(member.displayName)) member.displayName = this.name;
+    public List<StatusActivity> fetchActivities() {
+        List<StatusActivity> statuses = new ArrayList<StatusActivity>();
+        
+        StringBuilder url = new StringBuilder("https://www.googleapis.com/plus/v1/people/")
+                .append(this.googleId)
+                .append("/activities/public?key=AIzaSyC4xOkQsEPJcUKUvQGL6T7RZkrIIxSuZAg");
+        JsonObject response = JSON.getAsObject(WS.url(url.toString()).get().getString());    // Unauthenticated
+        JsonArray activities = response.get("items").getAsJsonArray();
+        DateFormat googleFormatter = new SimpleDateFormat(DATE_FORMAT, Locale.ENGLISH);
+        for (JsonElement element : activities) {
+            JsonObject activity = element.getAsJsonObject();
+            try {
+                String content = activity.get("object").getAsJsonObject().get("content").getAsString();
+                Date date = googleFormatter.parse(activity.get("published").getAsString());
+                String statusId = activity.get("id").getAsString();
+                String statusUrl = activity.get("url").getAsString();
+                statuses.add(new StatusActivity(this.member, date, this.provider, content, statusUrl, statusId));
+            } catch (ParseException pe) {
+                Logger.error("ouch! parse exception " + pe.getMessage());
+            }
         }
-    }
+        return statuses;
+     }
 
-    @Override
-    public Member findCorrespondingMember() {
-        return Member.find("byEmail", getOAuthLogin()).first();
+    public void enhance(Collection<StatusActivity> activities) {
+        // TODO Google enhance
     }
 }
