+++ app/models/OAuthAccount.java
@@ -2,27 +2,32 @@ package models;
 
 import javax.persistence.Entity;
 import play.data.validation.Required;
+import play.db.jpa.Model;
 
 /**
  * An OAuth account on a third-party authentication provider
  * @author Sryl <cyril.lacote@gmail.com>
  */
 @Entity
-public abstract class OAuthAccount extends Account {
+public abstract class OAuthAccount extends Model {
 
     @Required
     public String token;
     @Required
     public String secret;
     
-    /** ID of last status retrieved */
-    public String lastStatusId;
-
     public OAuthAccount(ProviderType provider, String token, String secret) {
-        super(provider);
+        super();
         this.token = token;
         this.secret = secret;
     }
+    
+    public abstract Account getAccount();
+    
+    /**
+     * Initialize member profile from account data
+     */
+    public abstract void initMemberProfile();
 
     /**
      * Template method to retrieve login value
@@ -31,8 +36,8 @@ public abstract class OAuthAccount extends Account {
     public abstract String getOAuthLogin();
 
     /**
-     * Find an existing member which could own this account
-     * @return Corresponding member found, null if not
+     * Find an existing account for this authentication account
+     * @return Corresponding account found, null if not
      */
     public abstract Member findCorrespondingMember();
 }
