+++ test/helpers/oauth/AbstractOAuthProviderImplTest.java
@@ -1,8 +1,5 @@
 package helpers.oauth;
 
-import helpers.oauth.OAuthProviderFactory;
-import helpers.oauth.AbstractOAuthProviderImpl;
-import helpers.oauth.OAuthProvider;
 import models.ProviderType;
 import org.apache.commons.lang.StringUtils;
 import org.junit.Test;
